Module zaino_fetch::chain::transaction

Transaction fetching and deserialization functionality.

Structs

Action

An Orchard action.

FullTransaction

Zingo-Indexer struct for a full zcash transaction.

JoinSplit

joinSplit is a JoinSplit description as described in 7.2 of the Zcash protocol spec. Its exact contents differ by transaction version and network upgrade level. Only version 4 is supported, no need for proofPHGR13.

Output

output is a Sapling Output Description as described in section 7.4 of the Zcash protocol spec.

Spend

spend is a Sapling Spend Description as described in 7.3 of the Zcash protocol specification.

TransactionData

Full Zcash Transactrion data.

TxIn

Txin format as described in https://en.bitcoin.it/wiki/Transaction

TxOut

Txout format as described in https://en.bitcoin.it/wiki/Transaction