Module zaino_fetch::chain::utils

Blockcache utility functionality.

Structs

CompactSize

Zcash CompactSize implementation taken from LibRustZcash::zcash_encoding to simplify dependency tree.

Constants

MAX_COMPACT_SIZE

The maximum allowed value representable as a [CompactSize]

Traits

ParseFromSlice

Used for decoding zcash blocks from a bytestring.

Functions

display_txids_to_server

Takes a vec of big endian hex encoded txids and returns them as a vec of little endian raw bytes.

read_bool

Reads the next byte from cursor into a bool, returns error message given if eof is reached..

read_bytes

Reads the next n bytes from cursor into a vec, returns error message given if eof is reached..

read_i32

Reads the next 4 bytes from cursor into an i32, returns error message given if eof is reached..

read_u32

Reads the next 4 bytes from cursor into a u32, returns error message given if eof is reached..

read_u64

Reads the next 8 bytes from cursor into a u64, returns error message given if eof is reached..

read_zcash_script_i64

Reads and interprets a Zcash (Bitcoin) custom compact integer encoding used for int64 numbers in scripts.

skip_bytes

Skips the next n bytes in cursor, returns error message given if eof is reached.