Module zaino_fetch::jsonrpc::response

Response types for jsonRPC client.

Structs

GetBalanceResponse

The transparent balance of a set of addresses.

GetBlockHash

Response to a getbestblockhash and getblockhash RPC request.

GetBlockTrees

Information about the sapling and orchard note commitment trees if any.

GetBlockchainInfoResponse

Response to a getblockchaininfo RPC request.

GetInfoResponse

Response to a getinfo RPC request.

GetSubtreesResponse

Contains the Sapling or Orchard pool label, the index of the first subtree in the list, and a list of subtree roots and end heights.

GetTreestateResponse

Contains the hex-encoded Sapling & Orchard note commitment trees, and their corresponding [block::Hash], [Height], and block time.

GetUtxosResponse

This is used for the output parameter of [JsonRpcConnector::get_address_utxos].

OrchardTrees

Orchard note commitment tree information.

SaplingTrees

Sapling note commitment tree information.

Script

Wrapper struct for a zebra Scrypt.

SendTransactionResponse

Contains the hex-encoded hash of the sent transaction.

SerializedBlock

A wrapper struct for a zebra serialized block.

SerializedTransaction

A wrapper struct for a zebra serialized transaction.

SubtreeRpcData

Wrapper struct for a zebra SubtreeRpcData.

TxidsResponse

Vec of transaction ids, as a JSON array.

Enums

GetBlockResponse

Contains the hex-encoded hash of the sent transaction.

GetTransactionResponse

Contains raw transaction, encoded as hex bytes.